+++ testsrc/ExperimentTest.cpp
@@ -33,9 +33,9 @@ int main(int argc, const char* argv[])
 
     vector<int> partIDs;
     exp.particleIDsMicrograph(partIDs, 0);
-    for (int i = 0; i < partIDs.size(); i++)
+    for (size_t i = 0; i < partIDs.size(); i++)
         cout << partIDs[i] << endl;
     exp.particleIDsGroup(partIDs, 0);
-    for (int i = 0; i < partIDs.size(); i++)
+    for (size_t i = 0; i < partIDs.size(); i++)
         cout << partIDs[i] << endl;
 }
